Crate atomic_cell

Structs

AtomicCell
A cell holding values protected by an atomic lock.

CleanGuard
A guard used to drop the value held by a CleanMeUp at a deterministic point in code. This is designed as an alternative to Drop for global variables.

StaticCell
A cell holding values protected by an atomic lock.

Traits

CleanMeUp
An object that needs to be cleaned up.